# packages/api-spec/generated/openapi.yaml
openapi: 3.1.0
info:
  title: Bytecrunch Contracts API
  version: 0.0.0
...
paths:
...
  /integration/v1/conditions/evaluate:
    post:
      operationId: MachineConditionOperations_evaluate
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ConditionEvaluation'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateConditions'
      security:
        - IntegrationOAuth2:
            - conditions:read
            - signing_sessions:write
  /integration/v1/signing-sessions:
    post:
      operationId: MachineSigningSessionOperations_create
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/IntegrationSessionResult'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIntegrationSession'
      security:
        - IntegrationOAuth2:
            - conditions:read
            - signing_sessions:write
  /oauth/token:
    post:
      operationId: IntegrationTokenOperations_issue
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/IntegrationTokenResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationTokenRequest'
...
components:
...
  schemas:
...
    AgreementCondition:
      type: object
      required:
        - kind
        - templateKey
      properties:
        kind:
          $ref: '#/components/schemas/AgreementConditionKind'
        templateKey:
          type: string
        minimumVersion:
          type: integer
          format: int32
    AgreementConditionKind:
      type: string
      enum:
        - subject_signed
        - agreement_executed
...
    AgreementStatus:
      type: string
      enum:
        - draft
        - in_review
        - ready_for_signature
        - out_for_signature
        - partially_signed
        - executed
        - declined
        - voided
        - expired
    ConditionEvaluation:
      type: object
      required:
        - decisionId
        - integrationKey
        - issuer
        - subject
        - met
        - operator
        - evaluatedAt
        - conditions
      properties:
        decisionId:
          type: string
        integrationKey:
          type: string
        issuer:
          type: string
        subject:
          type: string
        met:
          type: boolean
        operator:
          $ref: '#/components/schemas/RequirementOperator'
        evaluatedAt:
          type: string
          format: date-time
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/ConditionResult'
    ConditionResult:
      type: object
      required:
        - kind
        - templateKey
        - minimumVersion
        - met
      properties:
        kind:
          $ref: '#/components/schemas/AgreementConditionKind'
        templateKey:
          type: string
        minimumVersion:
          type: integer
          format: int32
        met:
          type: boolean
        status:
          $ref: '#/components/schemas/AgreementStatus'
        agreementId:
          type: string
        templateVersion:
          type: integer
          format: int32
        signedAt:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        executedAt:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
...
    CreateIntegrationSession:
      type: object
      required:
        - integrationKey
        - subject
        - email
        - templateKey
        - returnUrl
      properties:
        integrationKey:
          type: string
        subject:
          type: string
        email:
          type: string
        displayName:
          type: string
        templateKey:
          type: string
        title:
          type: string
        returnUrl:
          type: string
          format: uri
        metadata:
          type: object
          unevaluatedProperties:
            type: string
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        issues:
          type: array
          items:
            type: string
    EvaluateConditions:
      type: object
      required:
        - integrationKey
        - subject
        - conditions
      properties:
        integrationKey:
          type: string
        subject:
          type: string
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/AgreementCondition'
        operator:
          $ref: '#/components/schemas/RequirementOperator'
...
    IntegrationSessionResult:
      type: object
      required:
        - agreementId
        - expiresAt
        - handoffUrl
      properties:
        agreementId:
          type: string
        expiresAt:
          type: string
          format: date-time
        handoffUrl:
          type: string
          format: uri
    IntegrationTokenRequest:
      type: object
      required:
        - grant_type
      properties:
        grant_type:
          type: string
          enum:
            - client_credentials
        scope:
          type: string
        client_id:
          type: string
        client_secret:
          type: string
    IntegrationTokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
        - scope
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum:
            - Bearer
        expires_in:
          type: integer
          format: int32
        scope:
          type: string
...
    RequirementOperator:
      type: string
      enum:
        - all
        - any
...
  securitySchemes:
    ContractsOAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: http://localhost:8080/realms/bytecrunch/protocol/openid-connect/token
          scopes:
            agreements:read: ''
            agreements:write: ''
            templates:read: ''
            templates:write: ''
            webhooks:manage: ''
    IntegrationOAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: http://localhost:3001/oauth/token
          scopes:
            conditions:read: ''
            signing_sessions:write: ''
servers:
  - url: http://localhost:3001
    description: Local development
    variables: {}